[FIX] stabilize test locations and token formats
