+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod agent;
 pub mod client;
 pub mod config;
+pub mod discovery;
 pub mod domain;
 pub mod error;
 pub mod extractor;
